# packages/core/docs/swagger.yaml
basePath: /
definitions:
  api.ErrorResponse:
    properties:
      error:
        example: 'image pull failed: not found'
        type: string
    type: object
  api.RunRequest:
    properties:
      cmd:
        example:
        - python3
        - -c
        - print('hello world')
        items:
          type: string
        minItems: 1
        type: array
      cwd:
        example: /app
        type: string
      env:
        example:
        - HOME=/root
        - PYTHONPATH=/app
        items:
          type: string
        type: array
      files:
        description: |-
          Files lists relative paths of files previously uploaded via POST /files.
          Each file is bind-mounted read-only at /<path> inside the container.
        example:
        - workspace/script.py
        - workspace/data.json
        items:
          type: string
        type: array
      image:
        example: python:3.12-slim
        type: string
      limits:
        $ref: '#/definitions/config.ResourceLimits'
      network:
        default: none
        description: |-
          Network sets the container network mode: none (default, no access),
          sandbox (isolated NAT namespace), or host (shared host network).
        enum:
        - none
        - sandbox
        - host
        example: none
        type: string
      persist:
        description: |-
          Persist keeps uploaded input files and captured output files after the run.
          Default false: all files are deleted once the response is returned.
        type: boolean
    required:
    - cmd
    - image
    type: object
  api.RunResponse:
    properties:
      exec_id:
        example: boxer-abc123
        type: string
      exit_code:
        example: 0
        type: integer
      stderr:
        example: ""
        type: string
      stdout:
        example: |
          hello world
        type: string
      wall_ms:
        example: 342
        type: integer
    type: object
  config.ResourceLimits:
    properties:
      cpu_cores:
        type: number
      memory_mb:
        type: integer
      nofile:
        type: integer
      pids_limit:
        type: integer
      wall_clock_secs:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: 'Sandbox execution service: pull any container image, run arbitrary
    commands inside gVisor.'
  license:
    name: MIT
  title: Boxer API
  version: "1.0"
paths:
  /files:
    get:
      description: Download any file by its relative path. To retrieve output files
        written by a container to /output/, use the path pattern output/<exec_id>/<filename>.
      parameters:
      - description: Relative file path (e.g. output/boxer-abc123/result.json)
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File contents
        "400":
          description: Missing or invalid path
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: File not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Download a file from the file store
      tags:
      - files
    post:
      consumes:
      - multipart/form-data
      description: Multipart upload. The stored file can be referenced by its path
        in POST /run — it is bind-mounted read-only at /<path> inside the container.
      parameters:
      - description: File content
        in: formData
        name: file
        required: true
        type: file
      - description: Relative destination path (e.g. workspace/script.py)
        in: formData
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Stored path
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Missing or invalid form fields
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "413":
          description: Upload exceeds configured limit
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Upload a file to the file store
      tags:
      - files
  /healthz:
    get:
      description: 'Returns {"ok": true} when the service is running'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Health check
      tags:
      - system
  /run:
    post:
      consumes:
      - application/json
      description: |-
        Pulls the image if not cached, constructs a hardened OCI bundle, and runs the command inside a gVisor sandbox.
        Files listed in `files` must be uploaded first via POST /files; each is bind-mounted read-only at /<path> inside the container.
        Output files written to /output/ inside the container are captured and retrievable via GET /files?path=output/<exec_id>/<filename>.
      parameters:
      - description: Execution parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.RunRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Command completed (any exit code is valid — check exit_code)
          schema:
            $ref: '#/definitions/api.RunResponse'
        "400":
          description: Invalid request body or file not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "408":
          description: Wall-clock timeout exceeded
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal error (image pull failed, runsc error)
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "507":
          description: stdout or stderr exceeded the configured output limit
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Execute a command in a sandboxed container
      tags:
      - execution
swagger: "2.0"